Re-factor run script for cmsweb services

+++ docker/crabserver/run.sh
@@ -1,27 +1,40 @@
 #!/bin/bash
+srv=`echo $USER | sed -e "s,_,,g"`
 
 # overwrite host PEM files in /data/srv area
 if [ -f /etc/secrets/robotkey.pem ]; then
-    sudo cp /etc/secrets/robotkey.pem /data/srv/current/auth/crabserver/dmwm-service-key.pem
-    sudo cp /etc/secrets/robotcert.pem /data/srv/current/auth/crabserver/dmwm-service-cert.pem
+    sudo cp /etc/secrets/robotkey.pem /data/srv/current/auth/$srv/dmwm-service-key.pem
+    sudo cp /etc/secrets/robotcert.pem /data/srv/current/auth/$srv/dmwm-service-cert.pem
+    sudo chown $USER.$USER /data/srv/current/auth/$srv/dmwm-service-key.pem
+    sudo chown $USER.$USER /data/srv/current/auth/$srv/dmwm-service-cert.pem
 fi
 
 # overwrite proxy if it is present in /etc/proxy
 if [ -f /etc/proxy/proxy ]; then
-    mkdir -p /data/srv/state/crabserver/proxy
-    ln -s /etc/proxy/proxy /data/srv/state/crabserver/proxy/proxy.cert
+    mkdir -p /data/srv/state/$srv/proxy
+    cp /etc/proxy/proxy /data/srv/state/$srv/proxy/proxy.cert
+    sudo chown $USER.$USER /data/srv/state/$srv/proxy/proxy.cert
     mkdir -p /data/srv/current/auth/proxy
-    ln -s /etc/proxy/proxy /data/srv/current/auth/proxy/proxy
+    cp /etc/proxy/proxy /data/srv/current/auth/proxy/proxy
+    sudo chown $USER.$USER /data/srv/current/auth/proxy/proxy
 fi
 
 # overwrite header-auth key file with one from secrets
 if [ -f /etc/secrets/hmac ]; then
-    sudo rm /data/srv/current/auth/wmcore-auth/header-auth-key
-    cp /etc/secrets/hmac /data/srv/current/auth/wmcore-auth/header-auth-key
+    if [ -f /data/srv/current/auth/wmcore-auth/header-auth-key ]; then
+        sudo rm /data/srv/current/auth/wmcore-auth/header-auth-key
+    fi
+    sudo cp /etc/secrets/hmac /data/srv/current/auth/wmcore-auth/header-auth-key
+    sudo chown $USER.$USER /data/srv/current/auth/wmcore-auth/header-auth-key
+    if [ -f /data/srv/current/auth/$srv/header-auth-key ]; then
+        sudo rm /data/srv/current/auth/$srv/header-auth-key
+    fi
+    sudo cp /etc/secrets/hmac /data/srv/current/auth/$srv/header-auth-key
+    sudo chown $USER.$USER /data/srv/current/auth/$srv/header-auth-key
 fi
 
 # use service configuration files from /etc/secrets if they are present
-cdir=/data/srv/current/config/crabserver
+cdir=/data/srv/current/config/$srv
 files=`ls $cdir`
 for fname in $files; do
     if [ -f /etc/secrets/$fname ]; then
@@ -32,15 +45,16 @@ for fname in $files; do
         sudo chown $USER.$USER $cdir/$fname
     fi
 done
-
-# add t0auth.py
-if [ -f /etc/secrets/CRABServerAuth.py ]; then
-    sudo rm /data/srv/current/auth/crabserver/CRABServerAuth.py
-    ln -s /etc/secrets/CRABServerAuth.py /data/srv/current/auth/crabserver/CRABServerAuth.py
-fi
+files=`ls /etc/secrets`
+for fname in $files; do
+    if [ ! -f $cdir/$fname ]; then
+        sudo cp /etc/secrets/$fname /data/srv/current/auth/$srv/$fname
+        sudo chown $USER.$USER /data/srv/current/auth/$srv/$fname
+    fi
+done
 
 # start the service
-/data/srv/current/config/crabserver/manage start 'I did read documentation'
+/data/srv/current/config/$srv/manage start 'I did read documentation'
 
 # run monitoring script
 if [ -f /data/monitor.sh ]; then
